Product API

+++ lib/Product API/dimensions_model_3.dart
@@ -1,7 +1,7 @@
 class DimensionsModel {
-  int depth;
-  int height;
-  int width;
+  double depth;
+  double height;
+  double width;
 
 
   DimensionsModel({
